+++ packages/iov-ethereum/src/httpjsonrpcclient.ts
@@ -1,6 +1,11 @@
 import axios from "axios";
 
-import { JsonRpcRequest, JsonRpcResponse, parseJsonRpcError, parseJsonRpcResponse } from "@iov/jsonrpc";
+import {
+  JsonRpcRequest,
+  JsonRpcResponse,
+  parseJsonRpcErrorResponse,
+  parseJsonRpcSuccessResponse,
+} from "@iov/jsonrpc";
 
 export class HttpJsonRpcClient {
   private readonly baseUrl: string;
@@ -13,12 +18,12 @@ export class HttpJsonRpcClient {
     const result = await axios.post(this.baseUrl, request);
     const responseBody = result.data;
 
-    const errorResponse = parseJsonRpcError(responseBody);
-    if (errorResponse) {
-      return errorResponse;
+    let response: JsonRpcResponse;
+    try {
+      response = parseJsonRpcErrorResponse(responseBody);
+    } catch (_) {
+      response = parseJsonRpcSuccessResponse(responseBody);
     }
-
-    const successResponse = parseJsonRpcResponse(responseBody);
-    return successResponse;
+    return response;
   }
 }